Run process_incoming locally, update "export" code
Instances launched using this script can now be used as either an HTTP server
node or a process_incoming node. The "telemetry-incoming" service now runs
locally (doesn't start a separate instance to do the work), so can be used in
an AutoScale group.

The "export" code changed to not require AWS credentials to be embedded in the
upstart scripts, so it can use IAM roles for authentication.