Clean up the output dir before beginning work.
If a previous run of the code was interrupted, leaving ".log"
files in the output directory, they would result in dupes when
processing is restarted.

This change removes any ".log" files in the output dir to make
sure this doesn't happen.  If you don't want to clean up first,
pass in the "--no-clean" parameter.